+++ internal/provider/data_source_meraki_network_group_policy.go
@@ -340,6 +340,10 @@ func (d *NetworkGroupPolicyDataSource) Schema(ctx context.Context, req datasourc
 				MarkdownDescription: "The ID of the vlan you want to tag. This only applies if `settings` is set to `custom`.",
 				Computed:            true,
 			},
+			"force_delete": schema.BoolAttribute{
+				MarkdownDescription: "If true, the system deletes the GP even if there are active clients using the GP. After deletion, active clients that were assigned to that Group Policy will be left without any policy applied. Default is false.",
+				Computed:            true,
+			},
 		},
 	}
 }