+++ internal/provider/model_meraki_network_group_policies.go
@@ -81,6 +81,7 @@ type NetworkGroupPoliciesItems struct {
 	SchedulingWednesdayTo                        types.String                                 `tfsdk:"scheduling_wednesday_to"`
 	VlanTaggingSettings                          types.String                                 `tfsdk:"vlan_tagging_settings"`
 	VlanTaggingVlanId                            types.String                                 `tfsdk:"vlan_tagging_vlan_id"`
+	ForceDelete                                  types.Bool                                   `tfsdk:"force_delete"`
 }
 
 type NetworkGroupPoliciesBonjourForwardingRules struct {