+++ internal/provider/model_meraki_network_group_policy.go
@@ -80,6 +80,7 @@ type NetworkGroupPolicy struct {
 	SchedulingWednesdayTo                        types.String                               `tfsdk:"scheduling_wednesday_to"`
 	VlanTaggingSettings                          types.String                               `tfsdk:"vlan_tagging_settings"`
 	VlanTaggingVlanId                            types.String                               `tfsdk:"vlan_tagging_vlan_id"`
+	ForceDelete                                  types.Bool                                 `tfsdk:"force_delete"`
 }
 
 type NetworkGroupPolicyBonjourForwardingRules struct {