+++ internal/provider/resource_meraki_network_group_policy_test.go
@@ -103,6 +103,9 @@ func TestAccMerakiNetworkGroupPolicy(t *testing.T) {
 		ImportStateVerifyIgnore: []string{"traffic_shaping_rules.0.priority"},
 		Check:                   resource.ComposeTestCheckFunc(checks...),
 	})
+	steps = append(steps, resource.TestStep{
+		Config: testAccMerakiNetworkGroupPolicyPrerequisitesConfig + testAccNetworkGroupPolicyConfigAdditional0,
+	})
 
 	resource.Test(t, resource.TestCase{
 		PreCheck:                 func() { testAccPreCheck(t) },
@@ -229,5 +232,13 @@ func testAccMerakiNetworkGroupPolicyConfig_all() string {
 // End of section. //template:end testAccConfigAll
 
 // Section below is generated&owned by "gen/generator.go". //template:begin testAccConfigAdditional
+
+const testAccNetworkGroupPolicyConfigAdditional0 = `
+resource "meraki_network_group_policy" "test" {
+  network_id   = meraki_network.test.id
+  name         = "No video streaming"
+  force_delete = true
+}
+`
 
 // End of section. //template:end testAccConfigAdditional