+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Add `meraki_network_alerts_settings` resource and data source
 - Fix issue with not handling paginated responses correctly, [link](https://github.com/CiscoDevNet/terraform-provider-meraki/issues/71)
+- Add `force_delete` attribute to `meraki_network_group_policy` resource, [link](https://github.com/CiscoDevNet/terraform-provider-meraki/issues/72)
 
 ## 1.0.0